// src/hooks/metric/protobuf/metrics.ts
export interface Player {
	/**
	 * @generated from protobuf field: string Version = 1 [json_name = "Version"];
	 */
	version: string;
}
/**
 * @generated from protobuf message proto.Session
 */
export interface Session {
	/**
	 * @generated from protobuf field: string ID = 1 [json_name = "ID"];
	 */
	iD: string;
	/**
	 * @generated from protobuf field: string ViewID = 2 [json_name = "ViewID"];
	 */
	viewID: string;
	/**
	 * @generated from protobuf field: string ExternalID = 3 [json_name = "ExternalID"];
	 */
	externalID: string;
}
/**
 * @generated from protobuf message proto.OS
 */
export interface OS {
	/**
	 * @generated from protobuf field: string Name = 1 [json_name = "Name"];
	 */
	name: string;
	/**
	 * @generated from protobuf field: string Version = 2 [json_name = "Version"];
	 */
	version: string;
}
/**
 * @generated from protobuf message proto.CPU
 */
export interface CPU {
	/**
	 * @generated from protobuf field: string Arch = 1 [json_name = "Arch"];
	 */
	arch: string;
	/**
	 * @generated from protobuf field: uint32 Num = 2 [json_name = "Num"];
	 */
	num: number;
}
/**
 * @generated from protobuf message proto.Device
 */
export interface Device {
	/**
	 * @generated from protobuf field: proto.CPU CPU = 1 [json_name = "CPU"];
	 */
	cPU?: CPU;
	/**
	 * @generated from protobuf field: string Type = 2 [json_name = "Type"];
	 */
	type: string;
	/**
	 * @generated from protobuf field: string Model = 3 [json_name = "Model"];
	 */
	model: string;
	/**
	 * @generated from protobuf field: string Vendor = 4 [json_name = "Vendor"];
	 */
	vendor: string;
	/**
	 * @generated from protobuf field: uint32 TotalMemory = 5 [json_name = "TotalMemory"];
	 */
	totalMemory: number;
}
/**
 * @generated from protobuf message proto.Browser
 */
export interface Browser {
	/**
	 * @generated from protobuf field: string Name = 1 [json_name = "Name"];
	 */
	name: string;
	/**
	 * @generated from protobuf field: string Version = 2 [json_name = "Version"];
	 */
	version: string;
	/**
	 * @generated from protobuf field: proto.BrowserEngine Engine = 3 [json_name = "Engine"];
	 */
	engine?: BrowserEngine;
}
/**
 * @generated from protobuf message proto.BrowserEngine
 */
export interface BrowserEngine {
	/**
	 * @generated from protobuf field: string Name = 1 [json_name = "Name"];
	 */
	name: string;
	/**
	 * @generated from protobuf field: string Version = 2 [json_name = "Version"];
	 */
	version: string;
}
/**
 * @generated from protobuf message proto.Screen
 */
export interface Screen {
	/**
	 * @generated from protobuf field: uint32 Width = 1 [json_name = "Width"];
	 */
	width: number;
	/**
	 * @generated from protobuf field: uint32 Height = 2 [json_name = "Height"];
	 */
	height: number;
	/**
	 * @generated from protobuf field: uint32 ColorDepth = 3 [json_name = "ColorDepth"];
	 */
	colorDepth: number;
}
/**
 * @generated from protobuf message proto.Timestamp
 */
export interface Timestamp {
	/**
	 * @generated from protobuf field: uint32 Client = 1 [json_name = "Client"];
	 */
	client: number;
	/**
	 * @generated from protobuf field: uint32 Server = 2 [json_name = "Server"];
	 */
	server: number;
}
/**
 * @generated from protobuf message proto.TechSpecs
 */
export interface TechSpecs {
	/**
	 * @generated from protobuf field: proto.OS OS = 1 [json_name = "OS"];
	 */
	oS?: OS;
	/**
	 * @generated from protobuf field: proto.Device Device = 2 [json_name = "Device"];
	 */
	device?: Device;
	/**
	 * @generated from protobuf field: proto.Screen Screen = 3 [json_name = "Screen"];
	 */
	screen?: Screen;
	/**
	 * @generated from protobuf field: proto.Browser Browser = 4 [json_name = "Browser"];
	 */
	browser?: Browser;
	/**
	 * @generated from protobuf field: proto.Timestamp Timestamp = 5 [json_name = "Timestamp"];
	 */
	timestamp?: Timestamp;
}
/**
 * @generated from protobuf message proto.Media
 */
export interface Media {
	/**
	 * @generated from protobuf field: string Type = 1 [json_name = "Type"];
	 */
	type: string; // vod/live
	/**
	 * @generated from protobuf field: string VideoID = 2 [json_name = "VideoID"];
	 */
	videoID: string;
	/**
	 * @generated from protobuf field: string FolderID = 3 [json_name = "FolderID"];
	 */
	folderID: string;
	/**
	 * @generated from protobuf field: string ProjectID = 4 [json_name = "ProjectID"];
	 */
	projectID: string;
	/**
	 * @generated from protobuf field: string WorkspaceID = 5 [json_name = "WorkspaceID"];
	 */
	workspaceID: string;
}
// ...
export interface ResourceTiming {
	/**
	 * @generated from protobuf field: float DNS = 1 [json_name = "DNS"];
	 */
	dNS: number;
	/**
	 * @generated from protobuf field: float FetchTime = 2 [json_name = "FetchTime"];
	 */
	fetchTime: number;
	/**
	 * @generated from protobuf field: float TTFB = 3 [json_name = "TTFB"];
	 */
	tTFB: number;
	/**
	 * @generated from protobuf field: float Download = 4 [json_name = "Download"];
	 */
	download: number;
	/**
	 * @generated from protobuf field: float Total = 5 [json_name = "Total"];
	 */
	total: number;
	/**
	 * @generated from protobuf field: uint32 Size = 6 [json_name = "Size"];
	 */
	size: number;
	/**
	 * @generated from protobuf field: string Source = 7 [json_name = "Source"];
	 */
	source: string;
}
/**
 * @generated from protobuf message proto.Performance
 */
export interface Performance {
	/**
	 * @generated from protobuf field: string NetEffectiveType = 1 [json_name = "NetEffectiveType"];
	 */
	netEffectiveType: string; // Network_Information_API
	/**
	 * @generated from protobuf field: proto.PerformanceMemory Memory = 2 [json_name = "Memory"];
	 */
	memory?: PerformanceMemory;
	/**
	 * @generated from protobuf field: proto.ResourceTiming ResourceTiming = 3 [json_name = "ResourceTiming"];
	 */
	resourceTiming?: ResourceTiming;
	/**
	 * @generated from protobuf field: uint32 Timestamp = 4 [json_name = "Timestamp"];
	 */
	timestamp: number;
}
/**
 * @generated from protobuf message proto.Event
 */
export interface Event {
	/**
	 * @generated from protobuf field: string Type = 1 [json_name = "Type"];
	 */
	type: string;
	/**
	 * @generated from protobuf field: float Value = 2 [json_name = "Value"];
	 */
	value: number;
	/**
	 * @generated from protobuf field: uint32 WatchedSec = 3 [json_name = "WatchedSec"];
	 */
	watchedSec: number;
	/**
	 * @generated from protobuf field: uint32 PreviewPos = 4 [json_name = "PreviewPos"];
	 */
	previewPos: number;
	/**
	 * @generated from protobuf field: uint32 CurrentPos = 5 [json_name = "CurrentPos"];
	 */
	currentPos: number;
	/**
	 * @generated from protobuf field: proto.Properties Properties = 6 [json_name = "Properties"];
	 */
	properties?: Properties;
	/**
	 * @generated from protobuf field: uint32 Duration = 7 [json_name = "Duration"];
	 */
	duration: number;
	/**
	 * @generated from protobuf field: uint32 Timestamp = 8 [json_name = "Timestamp"];
	 */
	timestamp: number;
}
// ...
export interface Error {
	/**
	 * @generated from protobuf field: string Type = 1 [json_name = "Type"];
	 */
	type: string;
	/**
	 * @generated from protobuf field: string Message = 2 [json_name = "Message"];
	 */
	message: string;
	/**
	 * @generated from protobuf field: uint32 Timestamp = 3 [json_name = "Timestamp"];
	 */
	timestamp: number;
}
/**
 * @generated from protobuf message proto.Batch
 */
export interface Batch {
	/**
	 * @generated from protobuf field: proto.Player Player = 1 [json_name = "Player"];
	 */
	player?: Player;
	/**
	 * @generated from protobuf field: proto.Media Media = 2 [json_name = "Media"];
	 */
	media?: Media;
	/**
	 * @generated from protobuf field: proto.Session Session = 3 [json_name = "Session"];
	 */
	session?: Session;
	/**
	 * @generated from protobuf field: proto.TechSpecs TechSpecs = 4 [json_name = "TechSpecs"];
	 */
	techSpecs?: TechSpecs;
	/**
	 * @generated from protobuf field: proto.ResourceTiming NavigationTiming = 5 [json_name = "NavigationTiming"];
	 */
	navigationTiming?: ResourceTiming;
	/**
	 * @generated from protobuf field: repeated proto.Error Errors = 6 [json_name = "Errors"];
	 */
	errors: Error[];
	/**
	 * @generated from protobuf field: repeated proto.Event Events = 7 [json_name = "Events"];
	 */
	events: Event[];
	/**
	 * @generated from protobuf field: repeated proto.Performance Performance = 8 [json_name = "Performance"];
	 */
	performance: Performance[];
}
/**
 * @generated from protobuf message proto.Container
 */
export interface Container {
	/**
	 * @generated from protobuf field: bytes Addr = 1 [json_name = "Addr"];
	 */
	addr: Uint8Array;
	/**
	 * @generated from protobuf field: proto.Batch Batch = 2 [json_name = "Batch"];
	 */
	batch?: Batch;
}
// @generated message type with reflection information, may provide speed optimized methods
class Player$Type extends MessageType<Player> {
	constructor() {
		super('proto.Player', [
			{
				no: 1,
				name: 'Version',
				kind: 'scalar',
				jsonName: 'Version',
				T: 9 /* ScalarType.STRING */,
				options: {'gogoproto.moretags': 'ch:"player_version"'},
			},
		]);
	}
}
/**
 * @generated MessageType for protobuf message proto.Player
 */
export const Player = new Player$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Session$Type extends MessageType<Session> {
	constructor() {
		super('proto.Session', [
			{no: 1, name: 'ID', kind: 'scalar', jsonName: 'ID', T: 9 /* ScalarType.STRING */},
			{no: 2, name: 'ViewID', kind: 'scalar', jsonName: 'ViewID', T: 9 /* ScalarType.STRING */},
			{
				no: 3,
				name: 'ExternalID',
				kind: 'scalar',
				jsonName: 'ExternalID',
				T: 9 /* ScalarType.STRING */,
			},
		]);
	}
}
/**
 * @generated MessageType for protobuf message proto.Session
 */
export const Session = new Session$Type();
// ...
export const OS = new OS$Type();
// ...
export const Device = new Device$Type();
// ...
export const Browser = new Browser$Type();
// ...
export const Screen = new Screen$Type();
// ...
export const Timestamp = new Timestamp$Type();
// @generated message type with reflection information, may provide speed optimized methods
class TechSpecs$Type extends MessageType<TechSpecs> {
	constructor() {
		super('proto.TechSpecs', [
			{
				no: 1,
				name: 'OS',
				kind: 'message',
				jsonName: 'OS',
				T: () => OS,
				options: {'gogoproto.nullable': false},
			},
			{
				no: 2,
				name: 'Device',
				kind: 'message',
				jsonName: 'Device',
				T: () => Device,
				options: {'gogoproto.nullable': false},
			},
			{
				no: 3,
				name: 'Screen',
				kind: 'message',
				jsonName: 'Screen',
				T: () => Screen,
				options: {'gogoproto.nullable': false},
			},
			{
				no: 4,
				name: 'Browser',
				kind: 'message',
				jsonName: 'Browser',
				T: () => Browser,
				options: {'gogoproto.nullable': false},
			},
			{
				no: 5,
				name: 'Timestamp',
				kind: 'message',
				jsonName: 'Timestamp',
				T: () => Timestamp,
				options: {'gogoproto.nullable': false},
			},
		]);
	}
}
/**
 * @generated MessageType for protobuf message proto.TechSpecs
 */
export const TechSpecs = new TechSpecs$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Media$Type extends MessageType<Media> {
	constructor() {
		super('proto.Media', [
			{no: 1, name: 'Type', kind: 'scalar', jsonName: 'Type', T: 9 /* ScalarType.STRING */},
			{no: 2, name: 'VideoID', kind: 'scalar', jsonName: 'VideoID', T: 9 /* ScalarType.STRING */},
			{
				no: 3,
				name: 'FolderID',
				kind: 'scalar',
				jsonName: 'FolderID',
				T: 9 /* ScalarType.STRING */,
			},
			{
				no: 4,
				name: 'ProjectID',
				kind: 'scalar',
				jsonName: 'ProjectID',
				T: 9 /* ScalarType.STRING */,
			},
			{
				no: 5,
				name: 'WorkspaceID',
				kind: 'scalar',
				jsonName: 'WorkspaceID',
				T: 9 /* ScalarType.STRING */,
			},
		]);
	}
}
/**
 * @generated MessageType for protobuf message proto.Media
 */
export const Media = new Media$Type();
// ...
export const PerformanceMemory = new PerformanceMemory$Type();
// ...
export const ResourceTiming = new ResourceTiming$Type();
// ...
export const Performance = new Performance$Type();
// ...
export const Event = new Event$Type();
// ...
export const Properties = new Properties$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Error$Type extends MessageType<Error> {
	constructor() {
		super('proto.Error', [
			{no: 1, name: 'Type', kind: 'scalar', jsonName: 'Type', T: 9 /* ScalarType.STRING */},
			{no: 2, name: 'Message', kind: 'scalar', jsonName: 'Message', T: 9 /* ScalarType.STRING */},
			{
				no: 3,
				name: 'Timestamp',
				kind: 'scalar',
				jsonName: 'Timestamp',
				T: 13 /* ScalarType.UINT32 */,
			},
		]);
	}
}
/**
 * @generated MessageType for protobuf message proto.Error
 */
export const Error = new Error$Type();
// @generated message type with reflection information, may provide speed optimized methods
class Batch$Type extends MessageType<Batch> {
	constructor() {
		super('proto.Batch', [
			{
				no: 1,
				name: 'Player',
				kind: 'message',
				jsonName: 'Player',
				T: () => Player,
				options: {'gogoproto.nullable': false},
			},
			{
				no: 2,
				name: 'Media',
				kind: 'message',
				jsonName: 'Media',
				T: () => Media,
				options: {'gogoproto.nullable': false},
			},
			{
				no: 3,
				name: 'Session',
				kind: 'message',
				jsonName: 'Session',
				T: () => Session,
				options: {'gogoproto.nullable': false},
			},
			{
				no: 4,
				name: 'TechSpecs',
				kind: 'message',
				jsonName: 'TechSpecs',
				T: () => TechSpecs,
				options: {'gogoproto.nullable': false},
			},
			{
				no: 5,
				name: 'NavigationTiming',
				kind: 'message',
				jsonName: 'NavigationTiming',
				T: () => ResourceTiming,
				options: {'gogoproto.nullable': false},
			},
			{
				no: 6,
				name: 'Errors',
				kind: 'message',
				jsonName: 'Errors',
				repeat: 1 /* RepeatType.PACKED */,
				T: () => Error,
				options: {'gogoproto.nullable': false},
			},
			{
				no: 7,
				name: 'Events',
				kind: 'message',
				jsonName: 'Events',
				repeat: 1 /* RepeatType.PACKED */,
				T: () => Event,
				options: {'gogoproto.nullable': false},
			},
			{
				no: 8,
				name: 'Performance',
				kind: 'message',
				jsonName: 'Performance',
				repeat: 1 /* RepeatType.PACKED */,
				T: () => Performance,
				options: {'gogoproto.nullable': false},
			},
		]);
	}
}
/**
 * @generated MessageType for protobuf message proto.Batch
 */
export const Batch = new Batch$Type();